Fix sass-loader syntax in webpack.config.js
sass-loader takes an options parameter. One of those options is
includePaths, an array of absolute paths to search thru for import
statements within .scss files.

includePaths must take an Array as its argument. Previously, we listed a
String as there was only one absolute path we wanted to search through.
This does not work.

This commit updates sass-loader.options to be an Array instead of a
String.